Fix example project analyzer errors (MA0048, S3903, CA2007)
Move EmployeeRecord and HighEarnerRecord to own files with namespace,
add ConfigureAwait(false) to all awaited tasks.
