+++ examples/Wolfgang.Etl.DbClient.Example/EmployeeRecord.cs
@@ -1,0 +1,9 @@
+namespace Wolfgang.Etl.DbClient.Example;
+
+public class EmployeeRecord
+{
+    public int Id { get; set; }
+    public string FirstName { get; set; } = string.Empty;
+    public string LastName { get; set; } = string.Empty;
+    public double Salary { get; set; }
+}
