+++ examples/Wolfgang.Etl.DbClient.Example/HighEarnerRecord.cs
@@ -1,0 +1,7 @@
+namespace Wolfgang.Etl.DbClient.Example;
+
+public class HighEarnerRecord
+{
+    public string FullName { get; set; } = string.Empty;
+    public double Salary { get; set; }
+}
